Add section two and three, add Rachel images
Lots done here, forgot to commit in smaller chunks. Wrote the html and css for sections two and three, and imported the images of Rachel that will be used throughout the file. There is likely a lot of excessive and/or unused code within style.css, so later commits will likely keep the display the same while cutting down on unnecessary code.